# c# - remove async machinery overheads (takes new external dependency)