--- a/src/csharp/Grpc.Core/Grpc.Core.csproj
+++ b/src/csharp/Grpc.Core/Grpc.Core.csproj
@@ -99,6 +99,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="PooledAwait" Version="1.0.25" />
     <PackageReference Include="System.Interactive.Async" Version="3.2.0" />
     <!-- System.Buffers *may* come in transitively, but: we can *always* use ArrayPool -->
     <PackageReference Include="System.Buffers" Version="4.5.0" />
@@ -110,9 +111,9 @@
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.5' ">
-    <PackageReference Include="System.Runtime.Loader" Version="4.0.0" />
-    <PackageReference Include="System.Threading.Thread" Version="4.0.0" />
-    <PackageReference Include="System.Threading.ThreadPool" Version="4.0.10" />
+    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
+    <PackageReference Include="System.Threading.Thread" Version="4.3.0" />
+    <PackageReference Include="System.Threading.ThreadPool" Version="4.3.0" />
   </ItemGroup>
 
   <Import Project="NativeDeps.csproj.include" />