--- a/src/csharp/Grpc.Core/Grpc.Core.csproj
+++ b/src/csharp/Grpc.Core/Grpc.Core.csproj
@@ -98,7 +98,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="PooledAwait" Version="1.0.31" />
+    <PackageReference Include="PooledAwait" Version="1.0.33" />
     <PackageReference Include="System.Interactive.Async" Version="3.2.0" />
     <PackageReference Include="System.Memory" Version="4.5.3" />
   </ItemGroup>