@mgravell 2019-08-23
Closing; it looks like the BCL folks are doing something very similar; that would be the obvious winner here