@jtattermusch 2019-08-26
> Closing; it looks like the BCL folks are doing something very similar; that would be the obvious winner here

Mind sharing a link for reference?