@mgravell 2019-08-26
Sure: https://github.com/dotnet/coreclr/pull/26310

In the week I'll do a second pass over the changes I had here, and see which - if any - can be lifted in isolation that aren't directly task machinery related.